Feature/#135 sub type

+++ src/main/java/com/studioreservation/domain/reservation/repository/custom/ReservationRepositoryCustom.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 
 import com.studioreservation.global.request.PageRequestDTO;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -13,7 +14,7 @@ public interface ReservationRepositoryCustom {
 
 	List<ReservedTimeResDTO> findReservedTime(Timestamp strtDt, Timestamp endDt, Long roomCd);
 
-	Integer sumTotalAmount(Timestamp strtDt, Timestamp endDt);
+	BigDecimal sumTotalAmount(Timestamp strtDt, Timestamp endDt);
 
 	ReservationStateResponse findCountByState(Timestamp strtDt, Timestamp endDt);
 }
